+++ spykfunc/dataio/morphologies.py
@@ -34,8 +34,28 @@ class MorphologyDB(object):
 
     @functools.lru_cache(None)
     def first_axon_section(self, morpho: str):
+        """Return distances for a synapse on the first axon segment.
+
+        To avoid closeness to the soma, shift the position by 0.5 μm or the
+        length of the first segment, whichever is less.
+
+        Returns a tuple with
+        * the index of the first axon section
+        * the offset of the center of the first segment on said section
+        * the fractional offset of this point
+        * the distance from the soma of this point
+        """
         types = self[morpho].section_types
-        return types.index(int(morphokit.SectionType.axon)) + 1
+        section_index = types.index(int(morphokit.SectionType.axon))
+        section = self[morpho].section(section_index)
+        section_length = section.pathlength(len(section.points) - 1)
+        section_distance = min(0.5, section.pathlength(1))
+        return (
+            section_index + 1,  # MorphoK does not include the soma!
+            section_distance,
+            section_distance / section_length,
+            section.distance_to_soma() + section_distance
+        )
 
     @functools.lru_cache(None)
     def soma_radius(self, morpho: str):