+++ CHANGES.rst
@@ -5,6 +5,10 @@ Changelog
 Version 0.15.7
 ==============
 
+Fixes:
+  - The `SynapseReposition` filter did not parse the recipe correctly. See
+    also FUNCZ-247_.
+
 Changes:
   - The SONATA input will now create the field `synapse_id`, hence
     deprecating the `AddID` filter.
@@ -254,3 +258,4 @@ First working version with 3 base filters:
 .. _FUNCZ-238: https://bbpteam.epfl.ch/project/issues/browse/FUNCZ-238
 .. _FUNCZ-242: https://bbpteam.epfl.ch/project/issues/browse/FUNCZ-242
 .. _FUNCZ-244: https://bbpteam.epfl.ch/project/issues/browse/FUNCZ-244
+.. _FUNCZ-245: https://bbpteam.epfl.ch/project/issues/browse/FUNCZ-247